fix scaling
