+++ turbopack/crates/turbo-tasks-backend/src/backend/mod.rs
@@ -1,5 +1,6 @@
 mod cell_data;
 mod counter_map;
+mod eviction;
 mod operation;
 mod snapshot_coordinator;
 mod storage;
@@ -52,7 +53,9 @@ use turbo_tasks::{
 use turbo_tasks::{FunctionId, TaskDirtyCause};
 use turbo_tasks_malloc::TurboMalloc;
 
+use self::eviction::EvictionControl;
 pub use self::{
+    eviction::EvictionMode,
     operation::AnyOperation,
     storage::{EvictionCounts, SpecificTaskDataCategory, TaskDataCategory},
 };
@@ -113,139 +116,6 @@ pub enum StorageMode {
     /// Queries the storage for cache entries that don't exist locally.
     /// On shutdown, pushes all changes to the backing storage.
     ReadWriteOnShutdown,
-}
-
-/// Strategy for evicting evictable tasks from in-memory storage after a
-/// snapshot. This is an EXPERIMENTAL FEATURE under development.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum EvictionMode {
-    /// Never evict.
-    Off,
-    /// Evict after a snapshot only once enough memory has been allocated since
-    /// the last eviction to justify the cost of restoring evicted tasks on
-    /// demand. Uses allocator statistics to estimate reclaimable memory.
-    /// See [`EvictionControl::auto_threshold_exceeded`].
-    Auto,
-    /// After every snapshot, evict all evictable tasks from memory, reloading
-    /// them from disk on demand.
-    Full,
-}
-
-/// Owns the eviction policy for the background snapshot loop: the configured
-/// [`EvictionMode`] plus the threshold bookkeeping for [`EvictionMode::Auto`].
-///
-/// The loop consults [`EvictionControl::should_evict`] once per snapshot cycle
-/// and calls [`EvictionControl::record_eviction`] after a sweep. All
-/// mode-specific decisions live here so the loop doesn't have to branch on the
-/// mode itself.
-///
-/// For [`EvictionMode::Auto`]: eviction reclaims memory but forces evicted tasks
-/// to be restored from disk on next access. A sweep is only worth that churn if
-/// it reclaims a meaningful amount of memory. We can't measure reclaimable memory
-/// directly, so we use the net bytes allocated since the last eviction (via
-/// [`TurboMalloc::memory_usage`]) as a proxy and require it to exceed a threshold.
-struct EvictionControl {
-    mode: EvictionMode,
-    /// Global net live bytes ([`TurboMalloc::memory_usage`]) sampled immediately
-    /// after the most recent eviction, or `None` before the first eviction.
-    /// Only meaningful in [`EvictionMode::Auto`].
-    memory_at_last_eviction: Option<usize>,
-}
-
-impl EvictionControl {
-    fn new(mode: EvictionMode) -> Self {
-        Self {
-            mode,
-            memory_at_last_eviction: None,
-        }
-    }
-
-    /// Whether any prior cycle has evicted. Derived from the recorded baseline,
-    /// which [`EvictionControl::record_eviction`] sets after each sweep.
-    fn has_evicted_before(&self) -> bool {
-        self.memory_at_last_eviction.is_some()
-    }
-
-    /// Whether to run an eviction sweep this snapshot cycle.
-    ///
-    /// `snapshot_had_new_data` is whether the just-completed snapshot persisted
-    /// new data. We only evict when there's new data to persist (the common case)
-    /// or on the very first eviction after startup (data was already on disk from
-    /// a prior run, so `snapshot_had_new_data` may be false but in-memory state
-    /// can still be evicted).
-    ///
-    /// Within that, `Off` never evicts, `Full` always evicts, and `Auto` requires
-    /// enough net memory allocated since the last eviction to justify the
-    /// restore-then-re-evict churn (always evicting the first time, since there's
-    /// no prior baseline).
-    fn should_evict(&self, snapshot_had_new_data: bool) -> bool {
-        // Only evict when there's new data to persist, or on the very first
-        // eviction after startup (restored on-disk state can be reclaimed even
-        // when this snapshot had no new data).
-        if !snapshot_had_new_data && self.has_evicted_before() {
-            return false;
-        }
-        match self.mode {
-            EvictionMode::Off => false,
-            EvictionMode::Full => true,
-            EvictionMode::Auto => self.auto_threshold_exceeded(),
-        }
-    }
-
-    /// For [`EvictionMode::Auto`]: whether enough net memory has been allocated
-    /// since the last eviction to justify another sweep. Always evicts the first
-    /// time (no prior baseline). The threshold scales down under OS memory
-    /// pressure so we evict more eagerly when memory is tight.
-    fn auto_threshold_exceeded(&self) -> bool {
-        /// Minimum net bytes ([`TurboMalloc::memory_usage`] delta) that must be
-        /// allocated since the last eviction before another is worthwhile.
-        /// Allocated bytes are the proxy for how much a sweep would reclaim.
-        /// Default 128 MiB; overridable via `TURBO_ENGINE_EVICT_MIN_BYTES`.
-        static MIN_EVICT_BYTES: LazyLock<usize> = LazyLock::new(|| {
-            std::env::var("TURBO_ENGINE_EVICT_MIN_BYTES")
-                .ok()
-                .and_then(|v| v.parse::<usize>().ok())
-                .unwrap_or(128 * 1024 * 1024)
-        });
-        /// At full pressure the threshold shrinks to `(1 - 0.75) = 25%` of base.
-        const MAX_PRESSURE_REDUCTION: f64 = 0.75;
-
-        let Some(last) = self.memory_at_last_eviction else {
-            // First eviction: no baseline to compare against, so always evict.
-            return true;
-        };
-        let current = TurboMalloc::memory_usage();
-        // saturating: if memory shrank since the last eviction there is nothing
-        // new to reclaim, so the delta is 0 and we skip.
-        let allocated_since = current.saturating_sub(last);
-        let threshold = scale_threshold(
-            *MIN_EVICT_BYTES,
-            TurboMalloc::memory_pressure(),
-            MAX_PRESSURE_REDUCTION,
-        );
-        allocated_since >= threshold
-    }
-
-    /// Record the post-eviction memory floor as the new baseline.
-    fn record_eviction(&mut self) {
-        self.memory_at_last_eviction = Some(TurboMalloc::memory_usage());
-    }
-}
-
-/// Scale a base eviction threshold down according to OS memory pressure.
-///
-/// `factor = 1.0 - (pressure / 100) * max_reduction`. At pressure 0 the base is
-/// unchanged; at pressure 100 it is reduced by at most `max_reduction`. The
-/// result never reaches zero, so we always require *some* allocation before
-/// evicting. When pressure is unavailable (`None`) the base is returned unchanged.
-fn scale_threshold(base: usize, pressure: Option<u8>, max_reduction: f64) -> usize {
-    match pressure {
-        Some(p) => {
-            let p = (p.min(100) as f64) / 100.0;
-            (base as f64 * (1.0 - p * max_reduction)) as usize
-        }
-        None => base,
-    }
 }
 
 pub struct BackendOptions {
@@ -3999,102 +3869,4 @@ fn encode_task_data(
             })?;
     }
     Ok(SmallVec::from_slice(scratch_buffer))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{EvictionControl, EvictionMode, scale_threshold};
-
-    const BASE: usize = 128 * 1024 * 1024;
-    const MAX_REDUCTION: f64 = 0.75;
-
-    #[test]
-    fn off_mode_never_evicts() {
-        let mut control = EvictionControl::new(EvictionMode::Off);
-        for &new_data in &[true, false] {
-            assert!(!control.should_evict(new_data));
-            // Even after a (forced) eviction, Off never evicts.
-            control.record_eviction();
-            assert!(!control.should_evict(new_data));
-        }
-    }
-
-    #[test]
-    fn full_mode_evicts_on_new_data() {
-        let mut control = EvictionControl::new(EvictionMode::Full);
-        // New data → always evict, before and after a prior eviction.
-        assert!(control.should_evict(true));
-        control.record_eviction();
-        assert!(control.should_evict(true));
-    }
-
-    #[test]
-    fn full_mode_evicts_first_time_without_new_data() {
-        let mut control = EvictionControl::new(EvictionMode::Full);
-        // No new data, but never evicted before → first eviction still runs.
-        assert!(control.should_evict(false));
-        // No new data and already evicted → skip.
-        control.record_eviction();
-        assert!(!control.should_evict(false));
-    }
-
-    #[test]
-    fn auto_mode_evicts_first_time() {
-        // Fresh control has no baseline, so the first eligible cycle always
-        // evicts regardless of the memory threshold.
-        let mut control = EvictionControl::new(EvictionMode::Auto);
-        assert!(control.should_evict(true));
-        assert!(control.should_evict(false));
-        // But still respects the new-data/first-time trigger once it has evicted.
-        control.record_eviction();
-        assert!(!control.should_evict(false));
-    }
-
-    #[test]
-    fn scale_threshold_unavailable_pressure_uses_base() {
-        assert_eq!(scale_threshold(BASE, None, MAX_REDUCTION), BASE);
-    }
-
-    #[test]
-    fn scale_threshold_zero_pressure_uses_base() {
-        assert_eq!(scale_threshold(BASE, Some(0), MAX_REDUCTION), BASE);
-    }
-
-    #[test]
-    fn scale_threshold_full_pressure_applies_max_reduction() {
-        // At pressure 100 the threshold is (1 - 0.75) = 25% of base.
-        assert_eq!(scale_threshold(BASE, Some(100), MAX_REDUCTION), BASE / 4);
-    }
-
-    #[test]
-    fn scale_threshold_clamps_pressure_over_100() {
-        // Pressure is documented as 0..=100; values above clamp to 100.
-        assert_eq!(
-            scale_threshold(BASE, Some(200), MAX_REDUCTION),
-            scale_threshold(BASE, Some(100), MAX_REDUCTION)
-        );
-    }
-
-    #[test]
-    fn scale_threshold_is_monotonic_non_increasing() {
-        let mut prev = scale_threshold(BASE, Some(0), MAX_REDUCTION);
-        for p in 1..=100u8 {
-            let cur = scale_threshold(BASE, Some(p), MAX_REDUCTION);
-            assert!(
-                cur <= prev,
-                "threshold should not increase with pressure: p={p}, cur={cur}, prev={prev}"
-            );
-            prev = cur;
-        }
-    }
-
-    #[test]
-    fn scale_threshold_never_zero_for_nonzero_base() {
-        for p in 0..=100u8 {
-            assert!(
-                scale_threshold(BASE, Some(p), MAX_REDUCTION) > 0,
-                "threshold reached zero at pressure {p}"
-            );
-        }
-    }
 }
